backend concluido e seguro

+++ backend/src/controllers/cardController.js
@@ -1,29 +1,78 @@
 const mongoose = require('mongoose');
 const cardModel = require('../models/cardModel');
+const userModel = require('../models/userModel');
+
+const userValidation = async (username, password) => {
+  const user = await userModel.findOne({ username });
+  if (user){
+    if (user.password === password){
+      return (true);
+    }
+  }
+  return (false)
+}
 
 module.exports = {
   async loadAllCards(req, res) {
-    const cards = await cardModel.find();
-    return res.json(cards);
+    const { username } = req.params;
+    const { password } = req.body;
+    const { label } = req.body;
+
+    const validation = await userValidation(username, password);
+
+    if (validation) {
+      let cards = ''
+
+      if (label !== ''){
+        cards = await cardModel.find({ username, label });
+      } else {
+        cards = await cardModel.find({ username });
+      }
+      return res.json(cards);
+    }
+    return res.send('404') // access denied
   },
 
   async createNewCard(req, res) {
-    const newCard = await cardModel.create(req.body);
+    const { username } = req.body.validationInfo;
+    const { password } = req.body.validationInfo;
 
-    return res.json(newCard);
+    const validation = await userValidation(username, password);
+
+    if (validation){
+      await cardModel.create(req.body.data);
+      return res.send('200'); // successfully created
+    }
+    return res.send('404'); // access denied
   },
 
   async updateCard(req, res){
-    const { _id } = req.params;
+    const { username } = req.body.validationInfo;
+    const { password } = req.body.validationInfo;
 
-    const card = await cardModel.findByIdAndUpdate(_id, req.body, { new: true });
+    const validation = await userValidation(username, password);
 
-    return res.json(card)
+    if (validation) {
+      const { _id } = req.params;
+
+      await cardModel.findByIdAndUpdate(_id, req.body.data, { new: true });
+      return res.send('200'); // successfully updated
+    }
+    return res.send('404'); // access denied
   },
 
   async deleteCard(req, res){
-    const { _id } = req.params;
+    const { username } = req.body;
+    const { password } = req.body;
 
-    await cardModel.findByIdAndRemove(_id);
+    const validation = await userValidation(username, password);
+
+    if (validation){
+      const { _id } = req.params;
+
+      await cardModel.findByIdAndRemove(_id);
+      return res.send('200') // successfully deleted
+    }
+    return res.send('404'); // access denied
   }
 };
